+++ src/components/admin/briefing-compliance/index.ts
@@ -1,0 +1,3 @@
+export { default as ComplianceRateTrend } from "./ComplianceRateTrend";
+export { default as TodayStatusBoard } from "./TodayStatusBoard";
+export { default as RepeatOffenders } from "./RepeatOffenders";